backend/requirements.txt
# SOVEREIGN 2030 — Backend Dependencies
# ================================================
# Standard: Production-grade, pinned versions

# Backend Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Validation & Settings
pydantic==2.9.0
pydantic-settings==2.5.2

# Firebase / Google Cloud
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.71.0
google-cloud-secret-manager==2.20.0

# Stripe Billing
stripe==10.12.0

# HTTP Client (finAPI, external APIs)
httpx==0.27.0

# AI & Rate Limiting
google-genai==0.3.0
slowapi==0.1.9
